# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: 'ToggleBox: Remote Config & Feature Flags API'
  description: |
    **ToggleBox** is a multi-platform remote configuration and feature flag management service.

    ## Key Features

    - **Remote Configuration**: Version-controlled, immutable config management
    - **Feature Flags**: Flexible toggles with percentage and targeted rollouts
    - **Multi-Platform**: Separate configs for web, mobile, and other platforms
    - **Environment Management**: Staging, production, and custom environments
    - **CDN Caching**: CloudFront-backed caching with cache invalidation
    - **Client SDKs**: JavaScript, Next.js, and Expo/React Native

    ## API Structure

    ### Public Endpoints (Read-Only)
    - **Path**: `/api/v1/platforms/*`
    - **Access**: Publicly accessible, no authentication
    - **Methods**: GET only
    - **Caching**: CloudFront CDN with configurable TTL

    ### Authentication Endpoints
    - **Paths**: `/api/v1/auth/*`, `/api/v1/users/*`, `/api/v1/api-keys/*`
    - **Access**: Public for registration/login, authenticated for management
    - **Methods**: POST, GET, PATCH, DELETE

    ### Internal Endpoints (Write Operations)
    - **Path**: `/api/v1/internal/*`
    - **Access**: Network-restricted or authenticated
    - **Methods**: POST, PUT, PATCH, DELETE
    - **Security**: JWT Bearer tokens or API Keys required

    ## Authentication

    **JWT Bearer Tokens** (user accounts):
    - Register: POST `/api/v1/auth/register`
...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /internal/cache/invalidate:
    post:
      tags:
        - Cache
      summary: Invalidate CloudFront cache
      description: Invalidate CDN cache with different granularity levels (global, platform, environment, or version).
      operationId: invalidateCache
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CacheInvalidation'
      responses:
        '200':
          description: Cache invalidation initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      invalidationId:
                        type: string
                      status:
                        type: string
                      paths:
                        type: array
                        items:
                          type: string
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /internal/webhook/cache/invalidations/{invalidationId}:
    get:
      tags:
        - Webhooks
      summary: Get invalidation status
      description: Check the status of a specific cache invalidation operation
      operationId: getInvalidationStatus
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: invalidationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invalidation status
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      status:
                        type: string
                        enum: [InProgress, Completed]
                      createTime:
                        type: string
                        format: date-time
                      paths:
                        type: array
                        items:
                          type: string
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /internal/webhook/cache/invalidations:
    get:
      tags:
        - Webhooks
      summary: List all invalidations
      description: List all recent cache invalidation operations
      operationId: listInvalidations
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: List of cache invalidations
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        status:
                          type: string
                          enum: [InProgress, Completed]
                        createTime:
                          type: string
                          format: date-time
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'

# ==============================================================================
# COMPONENTS
# ==============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication
...
  schemas:
...
    CacheInvalidation:
      type: object
      description: Cache invalidation request with different granularity levels.
      properties:
        platform:
          type: string
          description: Invalidate all caches for this platform
        environment:
          type: string
          description: Invalidate all caches for this environment (requires platform)
        version:
          type: string
          pattern: '^\d+\.\d+\.\d+$'
          description: Semantic version label to invalidate (requires platform + environment)
        global:
          type: boolean
          description: Invalidate all caches globally
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - timestamp
      properties:
        success:
          type: boolean
        error:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            type: string
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...